chore(css): remove `@modern-js/style-compiler` package
