+++ .changeset/strange-bikes-tap.md
@@ -1,0 +1,12 @@
+---
+'@modern-js/core': minor
+'@modern-js/css-config': minor
+'@modern-js/plugin-less': minor
+'@modern-js/plugin-sass': minor
+'@modern-js/plugin-unbundle': minor
+'@modern-js/module-tools': minor
+---
+
+chore(css): remove `@modern-js/style-compiler` package
+
+chore(css): 移除 `@modern-js/style-compiler`
