+++ packages/cli/css-config/src/index.ts
@@ -1,3 +1,1 @@
 export * from './postcss';
-export * from './less';
-export * from './sass';
