+++ packages/cli/plugin-less/package.json
@@ -51,7 +51,6 @@
   },
   "dependencies": {
     "@babel/runtime": "^7.18.0",
-    "@modern-js/css-config": "workspace:^1.2.8",
     "@modern-js/utils": "workspace:^1.7.12",
     "@types/less": "^3.0.3",
     "less": "^4.1.1",
@@ -60,7 +59,6 @@
   },
   "devDependencies": {
     "@modern-js/core": "workspace:*",
-    "@modern-js/style-compiler": "workspace:*",
     "@scripts/build": "workspace:*",
     "@scripts/jest-config": "workspace:*",
     "@types/jest": "^27",
