+++ packages/cli/plugin-sass/src/cli.ts
@@ -1,5 +1,7 @@
-import { Import, PLUGIN_SCHEMAS } from '@modern-js/utils';
+import { PLUGIN_SCHEMAS } from '@modern-js/utils';
 import type { CliPlugin } from '@modern-js/core';
+import { getSassLoaderOptions } from './options';
+import { moduleSassConfig } from './module-sass-config';
 
 const SASS_REGEX = /\.s(a|c)ss$/;
 const SASS_MODULE_REGEX = /\.module\.s(a|c)ss$/;
@@ -15,15 +17,6 @@ export default (): CliPlugin => ({
   name: '@modern-js/plugin-sass',
 
   setup: api => {
-    const cssConfig: typeof import('@modern-js/css-config') = Import.lazy(
-      '@modern-js/css-config',
-      require,
-    );
-    const msc: typeof import('./module-sass-config') = Import.lazy(
-      './module-sass-config',
-      require,
-    );
-
     return {
       validateSchema() {
         return PLUGIN_SCHEMAS['@modern-js/plugin-sass'];
@@ -40,7 +33,7 @@ export default (): CliPlugin => ({
               } = resolvedConfig;
 
               const { options, excludes } =
-                cssConfig.getSassLoaderOptions(resolvedConfig);
+                getSassLoaderOptions(resolvedConfig);
 
               const loaders = chain.module.rule(RULE.LOADERS);
 
@@ -88,7 +81,7 @@ export default (): CliPlugin => ({
           },
         };
       },
-      moduleSassConfig: msc.moduleSassConfig,
+      moduleSassConfig,
     };
   },
 });
