+++ packages/cli/plugin-sass/src/module-sass-config.ts
@@ -1,12 +1,11 @@
-import { getSassLoaderOptions } from '@modern-js/css-config';
-import type { NormalizedConfig } from '@modern-js/core';
-import { SassOptions as ResolvedSassOption } from '@modern-js/style-compiler';
+import type { NormalizedConfig, SassOption } from '@modern-js/core';
+import { getSassLoaderOptions } from './options';
 
 export const moduleSassConfig = ({
   modernConfig,
 }: {
   modernConfig: NormalizedConfig;
-}): ResolvedSassOption<'sync'> => {
+}): SassOption => {
   const { options } = getSassLoaderOptions(modernConfig);
   return {
     ...(options.sassOptions || { file: '' }),
