+++ packages/cli/plugin-unbundle/package.json
@@ -47,7 +47,7 @@
     "@babel/types": "^7.18.0",
     "@koa/cors": "^3.1.0",
     "@modern-js/create-request": "workspace:^1.3.0",
-    "@modern-js/css-config": "workspace:^1.2.8",
+    "@modern-js/css-config": "1.2.8",
     "@modern-js/esmpack": "workspace:^1.3.3",
     "@modern-js/plugin": "workspace:^1.4.2",
     "@modern-js/server": "workspace:^1.5.1",
