+++ packages/solutions/module-tools/package.json
@@ -71,12 +71,14 @@
     "@modern-js/plugin-changeset": "workspace:^1.4.0",
     "@modern-js/plugin-i18n": "workspace:^1.3.0",
     "@modern-js/plugin-jarvis": "workspace:^1.2.14",
-    "@modern-js/style-compiler": "workspace:^1.2.13",
     "@modern-js/utils": "workspace:^1.7.12",
     "@rollup/plugin-json": "~4.1.0",
     "@speedy-js/speedy-types": "0.13.2-alpha.3",
     "@speedy-js/speedy-core": "0.13.2-alpha.3",
     "@speedy-js/speedy-plugin-es5": "0.13.2-alpha.3",
+    "sass": "^1.45.0",
+    "postcss": "^8.4.14",
+    "postcss-import": "^14.0.2",
     "normalize-path": "^3.0.0",
     "p-map": "^4",
     "process.argv": "^0.6.0",
@@ -99,6 +101,7 @@
     "@types/react": "^17",
     "@types/react-dom": "^17",
     "@types/signal-exit": "^3.0.1",
+    "@types/postcss-import": "^12.0.1",
     "jest": "^27",
     "typescript": "^4",
     "ajv": "^8",
