+++ packages/solutions/module-tools/src/style-compiler/types/less-plugin-npm-import.d.ts
