+++ packages/solutions/module-tools/src/style-compiler/utils/cssRule.ts
