+++ packages/solutions/module-tools/src/style-compiler/utils/file.ts
