--- packages/toolkit/compiler/style/.npmignore
@@ -1,47 +1,0 @@
-.DS_Store
-
-.pnp
-.pnp.js
-.env.local
-.env.*.local
-.history
-.rts*
-*.log*
-*.pid
-*.pid.*
-*.report
-*.lcov
-lib-cov
-
-node_modules/
-.npm
-.lock-wscript
-.yarn-integrity
-.node_repl_history
-.nyc_output
-*.tsbuildinfo
-.eslintcache
-.sonarlint
-
-coverage/
-release/
-output/
-output_resource/
-tests/
-
-.vscode/**/*
-!.vscode/settings.json
-!.vscode/extensions.json
-.idea/
-
-**/*/api/typings/auto-generated
-src/
-
-modern.config.js
-modern.config.ts
-jest.config.js
-.eslintrc.js
-.eslintrc
-tsconfig.json
-
-.wireit/
