--- packages/toolkit/compiler/style/README.md
@@ -1,30 +1,0 @@
-
-<p align="center">
-  <a href="https://modernjs.dev" target="blank"><img src="https://lf3-static.bytednsdoc.com/obj/eden-cn/ylaelkeh7nuhfnuhf/modernjs-cover.png" width="300" alt="Modern.js Logo" /></a>
-</p>
-<p align="center">
-现代 Web 工程体系
-  <br/>
-  <a href="https://modernjs.dev" target="blank">
-    modernjs.dev
-  </a>
-</p>
-<p align="center">
-  The meta-framework suite designed from scratch for frontend-focused modern web development
-</p>
-
-# Introduction
-
-> The doc site ([modernjs.dev](https://modernjs.dev)) and articles are only available in Chinese for now, we are planning to add English versions soon.
-
-- [Modern.js: Hello, World!](https://zhuanlan.zhihu.com/p/426707646)
-
-## Getting Started
-
-- [Quick Start](https://modernjs.dev/docs/start)
-- [Guides](https://modernjs.dev/docs/guides)
-- [API References](https://modernjs.dev/docs/apis)
-
-## Contributing
-
-- [Contributing Guide](https://github.com/modern-js-dev/modern.js/blob/main/CONTRIBUTING.md)
