--- packages/toolkit/compiler/style/package.json
@@ -1,91 +1,0 @@
-{
-  "name": "@modern-js/style-compiler",
-  "description": "The meta-framework suite designed from scratch for frontend-focused modern web development.",
-  "homepage": "https://modernjs.dev",
-  "bugs": "https://github.com/modern-js-dev/modern.js/issues",
-  "repository": "modern-js-dev/modern.js",
-  "license": "MIT",
-  "keywords": [
-    "react",
-    "framework",
-    "modern",
-    "modern.js"
-  ],
-  "version": "1.2.14",
-  "jsnext:source": "./src/index.ts",
-  "types": "./src/index.ts",
-  "main": "./dist/js/node/index.js",
-  "module": "./dist/js/treeshaking/index.js",
-  "jsnext:modern": "./dist/js/modern/index.js",
-  "exports": {
-    ".": {
-      "node": {
-        "jsnext:source": "./src/index.ts",
-        "import": "./dist/js/modern/index.js",
-        "require": "./dist/js/node/index.js"
-      },
-      "default": "./dist/js/treeshaking/index.js"
-    }
-  },
-  "scripts": {
-    "prepare": "pnpm build",
-    "prepublishOnly": "only-allow-pnpm && pnpm build -- --platform",
-    "new": "modern new",
-    "build": "wireit",
-    "test": "wireit"
-  },
-  "dependencies": {
-    "@babel/runtime": "^7.18.0",
-    "@modern-js/utils": "workspace:^1.7.11",
-    "less": "^4.1.1",
-    "postcss": "^8.4.14",
-    "postcss-import": "^14.0.2",
-    "sass": "^1.45.0"
-  },
-  "devDependencies": {
-    "@types/jest": "^27",
-    "@types/less": "^3.0.3",
-    "@types/node": "^14",
-    "@types/postcss-import": "^12.0.1",
-    "@types/react": "^17",
-    "@types/react-dom": "^17",
-    "@types/tailwindcss": "^2.2.1",
-    "typescript": "^4",
-    "@scripts/build": "workspace:*",
-    "jest": "^27",
-    "@scripts/jest-config": "workspace:*"
-  },
-  "modernConfig": {
-    "output": {
-      "packageMode": "node-js"
-    }
-  },
-  "publishConfig": {
-    "registry": "https://registry.npmjs.org/",
-    "access": "public",
-    "types": "./dist/types/index.d.ts"
-  },
-  "wireit": {
-    "build": {
-      "command": "modern build",
-      "files": [
-        "src/**/*",
-        "tsconfig.json",
-        "package.json"
-      ],
-      "output": [
-        "dist/**/*"
-      ]
-    },
-    "test": {
-      "command": "jest --passWithNoTests",
-      "files": [
-        "src/**/*",
-        "tsconfig.json",
-        "package.json",
-        "tests/**/*"
-      ],
-      "output": []
-    }
-  }
-}
