--- packages/toolkit/compiler/style/tests/index.test.ts
@@ -1,7 +1,0 @@
-import { styleCompiler } from '../src';
-
-describe('@modern-js/style-compiler', () => {
-  it('default', () => {
-    expect(styleCompiler).toBeDefined();
-  });
-});
