--- packages/toolkit/compiler/style/tests/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "@modern-js/tsconfig/base",
-  "compilerOptions": {
-    "declaration": true,
-    "jsx": "preserve",
-    "baseUrl": "./",
-    "outDir": "./out",
-    "emitDeclarationOnly": true,
-    "isolatedModules": true,
-    "paths": {},
-    "types": ["node", "jest"]
-  }
-}
