--- packages/toolkit/compiler/style/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "@modern-js/tsconfig/base",
-  "compilerOptions": {
-    "declaration": false,
-    "jsx": "preserve",
-    "baseUrl": "./",
-    "isolatedModules": true,
-    "paths": {}
-  },
-  "include": ["src"]
-}
